Move service config checking into its own file
