+++ cmd/launcher/svc_windows.go
@@ -21,7 +21,6 @@ import (
 	"github.com/kolide/launcher/pkg/log/locallogger"
 	"github.com/kolide/launcher/pkg/log/teelogger"
 
-	"golang.org/x/sys/windows/registry"
 	"golang.org/x/sys/windows/svc"
 	"golang.org/x/sys/windows/svc/debug"
 )
@@ -79,10 +78,8 @@ func runWindowsSvc(args []string) error {
 		)
 	}()
 
-	// For Kolide installations, confirm that service configuration is up-to-date
-	if opts.KolideServerURL == "k2device.kolide.com" || opts.KolideServerURL == "k2device-preprod.kolide.com" {
-		checkServiceConfiguration(logger)
-	}
+	// Confirm that service configuration is up-to-date
+	checkServiceConfiguration(logger, opts)
 
 	level.Info(logger).Log(
 		"msg", "launching service",
@@ -194,76 +191,5 @@ func (w *winSvc) Execute(args []string, r <-chan svc.ChangeRequest, changes chan
 				level.Info(w.logger).Log("err", "unexpected control request", "control_request", c)
 			}
 		}
-	}
-}
-
-const (
-	launcherServiceRegistryKeyName = `SYSTEM\CurrentControlSet\Services\LauncherKolideK2Svc`
-
-	// DelayedAutostart is type REG_DWORD, i.e. uint32. We want to turn off delayed autostart.
-	delayedAutostartName            = `DelayedAutostart`
-	delayedAutostartDisabled uint32 = 0
-
-	// DependOnService is type REG_MULTI_SZ, i.e. a list of strings
-	dependOnServiceName = `DependOnService`
-	dnscacheService     = `Dnscache`
-)
-
-func checkServiceConfiguration(logger log.Logger) {
-	// Get launcher service key
-	launcherServiceKey, err := registry.OpenKey(registry.LOCAL_MACHINE, launcherServiceRegistryKeyName, registry.ALL_ACCESS)
-	if err != nil {
-		level.Error(logger).Log("msg", "could not open registry key", "key_name", launcherServiceRegistryKeyName, "err", err)
-		return
-	}
-
-	// Close it once we're done
-	defer func() {
-		if err := launcherServiceKey.Close(); err != nil {
-			level.Error(logger).Log("msg", "could not close registry key", "key_name", launcherServiceRegistryKeyName, "err", err)
-		}
-	}()
-
-	// Check to see if we need to turn off delayed autostart
-	currentDelayedAutostart, _, getDelayedAutostartErr := launcherServiceKey.GetIntegerValue(delayedAutostartName)
-	if getDelayedAutostartErr == nil && currentDelayedAutostart != uint64(delayedAutostartDisabled) {
-		// Turn off delayed autostart
-		if err := launcherServiceKey.SetDWordValue(delayedAutostartName, delayedAutostartDisabled); err != nil {
-			level.Error(logger).Log("msg", "could not set dword value for DelayedAutostart", "err", err)
-		}
-	}
-
-	// Check to see if we need to update the service to depend on Dnscache
-	serviceList, _, getServiceListErr := launcherServiceKey.GetStringsValue(dependOnServiceName)
-	if getServiceListErr != nil {
-		// If DependsOn isn't set at all yet, set it with the Dnscache service
-		if getServiceListErr.Error() == "The system cannot find the file specified." {
-			if err := launcherServiceKey.SetStringsValue(dependOnServiceName, []string{dnscacheService}); err != nil {
-				level.Error(logger).Log("msg", "could not set strings value for DependOnService", "err", err)
-			}
-			return
-		}
-
-		// In any other case, if we can't get the current value, we don't want to proceed --
-		// we don't want to wipe any current data from the list.
-		return
-	}
-
-	foundDnscacheInList := false
-	for _, service := range serviceList {
-		if service == dnscacheService {
-			foundDnscacheInList = true
-			break
-		}
-	}
-
-	if foundDnscacheInList {
-		return
-	}
-
-	// Set service to depend on Dnscache
-	serviceList = append(serviceList, dnscacheService)
-	if err := launcherServiceKey.SetStringsValue(dependOnServiceName, serviceList); err != nil {
-		level.Error(logger).Log("msg", "could not set strings value for DependOnService", "err", err)
 	}
 }
